Organised the classes
