+++ src/main/java/snap/Main.java
@@ -1,5 +1,7 @@
 package snap;
 
+import snap.gameLogic.GameController;
+
 public class Main {
     public static void main(String[] args) {
         GameController gameController = new GameController();
