+++ src/main/java/snap/gameLogic/CardGame.java
@@ -1,4 +1,6 @@
-package snap;
+package snap.gameLogic;
+
+import snap.gameLogic.Card;
 
 import java.util.*;
