+++ src/main/java/snap/player/Player.java
@@ -1,4 +1,4 @@
-package snap;
+package snap.player;
 
 public class Player {
